refactor the feature table parsers/serializers to _sequence_feature_vocabulary.py
